+++ src/main/java/com/newstoss/news/adapter/in/web/news/dto/v2/GetAllNewsDTO.java
@@ -1,6 +1,5 @@
 package com.newstoss.news.adapter.in.web.news.dto.v2;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,8 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class RelatedStockDTOv2 {
-    @JsonProperty("stock_id")
-    private String stockId;
-    private String stockName;
+public class GetAllNewsDTO {
+    private Integer skip;
+    private Integer limit;
 }
