+++ src/main/java/com/newstoss/news/adapter/in/web/news/dto/v2/Meta/NewsMetaDataDTO.java
@@ -1,8 +1,5 @@
-package com.newstoss.news.adapter.in.web.news.dto.v2;
+package com.newstoss.news.adapter.in.web.news.dto.v2.Meta;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.newstoss.news.adapter.out.news.dto.v2.MLIndustryListDTO;
-import com.newstoss.news.adapter.out.news.dto.v2.MLRelatedStockDTOv2;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,7 +9,7 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class NewsMetaDataDTO {
+public class    NewsMetaDataDTO {
     private String newsId;
     private String summary;
     private List<RelatedStockDTOv2> stockList;
