+++ src/main/java/com/newstoss/news/adapter/in/web/news/dto/v2/StockNewsDTO.java
@@ -1,0 +1,16 @@
+package com.newstoss.news.adapter.in.web.news.dto.v2;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.criteria.CriteriaBuilder;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class StockNewsDTO {
+    private Integer skip;
+    private Integer limit;
+    private String stockCode;
+}
