+++ src/main/java/com/newstoss/news/application/news/v1/impl/GetAllNewsV1.java
@@ -1,6 +1,6 @@
 package com.newstoss.news.application.news.v1.impl;
 
-import com.newstoss.news.adapter.in.web.news.dto.common.GetAllNewsDTO;
+import com.newstoss.news.adapter.in.web.news.dto.v2.GetAllNewsDTO;
 import com.newstoss.news.adapter.in.web.news.dto.v1.NewsDTOv1;
 import com.newstoss.news.adapter.out.news.dto.v1.MLNewsDTOv1;
 import com.newstoss.news.application.news.v1.port.in.GetAllNewsUseCase;
