+++ src/main/java/com/newstoss/news/application/news/v1/port/in/GetAllNewsUseCase.java
@@ -1,6 +1,6 @@
 package com.newstoss.news.application.news.v1.port.in;
 
-import com.newstoss.news.adapter.in.web.news.dto.common.GetAllNewsDTO;
+import com.newstoss.news.adapter.in.web.news.dto.v2.GetAllNewsDTO;
 import com.newstoss.news.adapter.in.web.news.dto.v1.NewsDTOv1;
 import java.util.List;
