+++ src/main/java/com/newstoss/news/application/news/v2/impl/MatchHighlightWithRelated.java
@@ -2,7 +2,7 @@ package com.newstoss.news.application.news.v2.impl;
 
 import com.newstoss.news.adapter.in.web.news.dto.v2.HighlightNewsDTO;
 import com.newstoss.news.adapter.in.web.news.dto.v2.NewsMathRelatedDTO;
-import com.newstoss.news.adapter.in.web.news.dto.v2.RelatedNewsDTOv2;
+import com.newstoss.news.adapter.in.web.news.dto.v2.Meta.RelatedNewsDTOv2;
 import com.newstoss.news.application.news.v2.port.in.GetRelatedNewsUseCaseV2;
 import com.newstoss.news.application.news.v2.port.in.MatchNewsWithRelatedUseCase;
 import lombok.RequiredArgsConstructor;
