+++ src/main/java/com/newstoss/news/application/news/v2/port/in/GetNewsMataDataUseCaseV2.java
@@ -1,6 +1,6 @@
 package com.newstoss.news.application.news.v2.port.in;
 
-import com.newstoss.news.adapter.in.web.news.dto.v2.NewsMetaDataDTO;
+import com.newstoss.news.adapter.in.web.news.dto.v2.Meta.NewsMetaDataDTO;
 
 public interface GetNewsMataDataUseCaseV2 {
     NewsMetaDataDTO exec(String newsId);
