+++ src/main/java/com/newstoss/news/application/news/v2/port/in/GetRelatedNewsUseCaseV2.java
@@ -2,7 +2,7 @@ package com.newstoss.news.application.news.v2.port.in;
 
 
 
-import com.newstoss.news.adapter.in.web.news.dto.v2.RelatedNewsDTOv2;
+import com.newstoss.news.adapter.in.web.news.dto.v2.Meta.RelatedNewsDTOv2;
 
 import java.util.List;
