+++ src/main/java/com/newstoss/news/application/redis/port/in/HighlightNewsCacheUseCase.java
@@ -1,6 +1,5 @@
 package com.newstoss.news.application.redis.port.in;
 
-import com.newstoss.news.adapter.in.web.news.dto.v1.NewsMathRelatedDTOTest;
 import com.newstoss.news.adapter.in.web.news.dto.v2.HighlightNewsDTO;
 import com.newstoss.news.adapter.in.web.news.dto.v2.NewsMathRelatedDTO;
